+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Conformance endpoint `/conformance` and root body `conformsTo` attribute.
+- Endpoint /product/{productId}/order-parameters.
+- Links in Product entity to order-parameters and constraints endpoints for that product.
+- Add links `opportunities` and `create-order` to Product
+- Add link `create-order` to OpportunityCollection
 
 ### Changed